requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
convert-to-queryset==0.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-crontab==0.7.1
django-filter==21.1
django-picklefield==3.1
django-q==1.3.9
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
